pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "PixelHand"
description = "An AI-powered tool for finding and managing duplicate and visually similar images."
version = "1.0.0"
readme = "README.md"
requires-python = ">=3.13"
license = "MIT"
authors = [{ name = "kilazz"}]

dependencies = [
    # Core Application & GUI
    "PySide6>=6.10.0",
    "send2trash>=1.8.3",
    # AI & Machine Learning
    "torch>=2.9.1",
    "transformers>=4.57.1",
    "sentencepiece>=0.2.1",
    "huggingface_hub[hf_xet]>=0.36.0", #transformers
    "onnx>=1.19.1",
    "onnxscript>=0.5.6",
    # Data Handling and Vector DB
    "lancedb>=0.25.3",
    "numpy>=2.3.4",
    "numba>=0.62.1",
    "pyarrow>=22.0.0",
    "polars>=1.35.2",
    "xxhash>=3.6.0",
    "ImageHash>=4.3.2",
    # Image Processing
    "simple-ocio>=0.2.0",
    "Pillow>=12.0.0",
    "OpenImageIO>=3.1.7.0",
    #"directxtex-decoder", #for DDS
]

[project.optional-dependencies]
# ONNX Runtime
cpu      = ["onnxruntime>=1.23.2"]
cuda     = ["onnxruntime-gpu>=1.23.2"]
directml = ["onnxruntime-directml>=1.23.0"]
webgpu   = ["onnxruntime-webgpu>=1.24.0.dev20251031004"]

dev = [
    "ruff",
    "snakeviz",
    "py-spy",
    "tabulate2",
    "psutil",
    "tqdm",
]

[tool.ruff]
exclude = [
	"__pycache__",
	"__pypackages__",
	".direnv",
	".eggs",
	".git",
	".git-rewrite",
	".hg",
	".mypy_cache",
	".nox",
	".pants.d",
	".pyenv",
	".pytype",
	".svn",
	".tox",
	"_build",
	"buck-out",
	"build",
	"dist",
	"site-packages",
    ".bzr",
    ".ipynb_checkpoints",
    ".pytest_cache",
    ".vscode",
    "node_modules",
	".ruff_cache",
	"_src_directxtex_decoder",
	"app_data",
	"venv",
	".venv",
	".venv-webgpu",
	".venv-directml",
	".venv-cuda",
	".venv-cpu",
]

line-length = 120
target-version = "py313"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle: errors (e.g., indentation, syntax)
    "W",   # pycodestyle: warnings (e.g., line length)
    "F",   # Pyflakes: logical errors (undefined names, unused imports)
    "I",   # isort: sorting and formatting of imports
    "B",   # flake8-bugbear: finds likely bugs and design problems
    "UP",  # pyupgrade: upgrades syntax to newer Python versions
    "SIM", # flake8-simplify: suggests simplifications for complex code
    "RUF", # Enables Ruff's native, fast, and often unique checks
]

ignore = ["E501"]
fixable = ["ALL"]
unfixable = []
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
exclude = ["app_data*"]
